move playtime to game detail
